pyproject.toml
[tool.poetry]
name = "test for shop"
version = "0.1.0"
description = ""
authors = ["Logunova Margarita <logunovar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selene = {version = "^2.0.0-beta.8", allow-prereleases = true}
Appium-Python-Client = "^2.6.1"
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
pytest-voluptuous = "^1.2.0"
voluptuous = ">=0.13.1,<0.14.0"
curlify = "^2.2.1"



[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
